Add main module and refactor
Pave the way for both tables and 3dgrid++ in one go
