+++ src/fmu/sumo/sim2sumo/tables.py
@@ -1,22 +1,21 @@
-"""Module for uploading tabular data from reservoir simulators to sumo
+"""Upload tabular data from reservoir simulators to sumo
    Does three things:
    1. Extracts data from simulator to arrow files
    2. Adds the required metadata while exporting to disc
    3. Uploads to Sumo
 """
 
-import argparse
 import logging
 import sys
-from pathlib import Path, PosixPath
+from pathlib import Path
 from typing import Union
 
 import pandas as pd
 import pyarrow as pa
 import res2df
 
-from ._special_treatments import SUBMOD_DICT, SUBMODULES, convert_options, tidy
-from .common import export_object, fix_suffix, upload, yaml_load
+from ._special_treatments import SUBMOD_DICT, convert_options, tidy
+from .common import export_object, fix_suffix, upload
 
 logging.getLogger(__name__).setLevel(logging.DEBUG)
 
@@ -30,10 +29,10 @@ SUBMOD_CONTENT.update(
 )
 
 
-def get_results(
+def get_table(
     datafile_path: str, submod: str, print_help=False, **kwargs
 ) -> Union[pa.Table, pd.DataFrame]:
-    """Fetch dataframe from simulator results
+    """Fetch arrow.table/pd.dataframe from simulator results
 
     Args:
         datafile_path (str): the path to the simulator datafile
@@ -43,7 +42,7 @@ def get_results(
     Returns:
         pd.DataFrame: the extracted data
     """
-    logger = logging.getLogger(__file__ + ".get_dataframe")
+    logger = logging.getLogger(__file__ + ".get_table")
     extract_df = SUBMOD_DICT[submod]["extract"]
     arrow = kwargs.get("arrow", True)
     datafile_path = fix_suffix(datafile_path)
@@ -80,8 +79,7 @@ def get_results(
                     )
                 except TypeError:
                     logger.warning("Type error, cannot convert to arrow")
-        except RuntimeError:
-            print(give_help(None))
+
         except TypeError:
             trace = sys.exc_info()[1]
         except FileNotFoundError:
@@ -97,10 +95,10 @@ def get_results(
     return output
 
 
-def export_results(
+def export_table(
     datafile_path: str,
     submod: str,
-    config_file="fmuconfig/output/global_variables.yml",
+    config,
     **kwargs,
 ) -> str:
     """Export csv file with specified datatype
@@ -112,248 +110,61 @@ def export_results(
     Returns:
         str: path of export
     """
-    logger = logging.getLogger(__file__ + ".export_results")
+    logger = logging.getLogger(__file__ + ".export_table")
     logger.debug("Export will be using these options: %s", kwargs)
-    frame = get_results(datafile_path, submod, **kwargs)
+    frame = get_table(datafile_path, submod, **kwargs)
 
     exp_path = export_object(
         datafile_path,
         submod,
-        config_file,
+        config,
         frame,
         SUBMOD_CONTENT.get(submod, "property"),
     )
     return exp_path
 
 
-def read_config(config, datafile=None, datatype=None):
-    """Read config settings
+def export_tables(sim2sumoconfig, config):
+    """Export from all datafiles and all submods selected
 
     Args:
-        config (dict): the settings for export of simulator results
-        kwargs (dict): overiding settings
+        datafiles (iterable): the datafiles to extract from
+        submods (iterable): the subtypes to extract
+        options (dict): the options to use
+        config (dict): the metadata
 
     Returns:
-        tuple: datafiles as list, submodules to use as list, and options as kwargs
+        tuple: path to where things have been exported, suffixes of exported files
     """
-    # datafile can be read as list, or string which can be either folder or filepath
-    logger = logging.getLogger(__file__ + ".read_config")
-    logger.debug("Input config keys are %s", config.keys())
-
-    simconfig = config.get("sim2sumo", {})
-    if isinstance(simconfig, bool):
-        simconfig = {}
-    datafiles = find_datafiles(datafile, simconfig)
-
-    submods = find_datatypes(datatype, simconfig)
-
-    options = simconfig.get("options", {"arrow": True})
-
-    logger.info(
-        "Running with: datafile(s): \n%s \n Types: \n %s \noptions:\n %s",
-        datafiles,
-        submods,
-        options,
-    )
-    return datafiles, submods, options
-
-
-def find_datatypes(datatype, simconfig):
-    """Find datatypes to extract
-
-    Args:
-        datatype (str or None): datatype to extract
-        simconfig (dict): the config file settings
-
-    Returns:
-        list or dict: data types to extract
-    """
-
-    if datatype is None:
-        submods = simconfig.get("datatypes", ["summary", "rft", "satfunc"])
-
-        if submods == "all":
-            submods = SUBMODULES
-    else:
-        submods = [datatype]
-    return submods
-
-
-def is_datafile(results: PosixPath) -> bool:
-    """Filter results based on suffix
-
-    Args:
-        results (PosixPath): path to file
-
-    Returns:
-        bool: true if correct suffix
-    """
-    valid = [".afi", ".DATA", ".in"]
-    return results.suffix in valid
-
-
-def find_datafiles(seedpoint, simconfig):
-    """Find all relevant paths that can be datafiles
-
-    Args:
-        seedpoint (str, list): path of datafile, or list of folders where one can find one
-        simconfig (dict): the sim2sumo config settings
-
-    Returns:
-        list: list of datafiles to interrogate
-    """
-
-    logger = logging.getLogger(__file__ + ".find_datafiles")
-    datafiles = []
-    seedpoint = simconfig.get("datafile", seedpoint)
-    if seedpoint is None:
-        datafiles = list(filter(is_datafile, Path().cwd().glob("*/model/*.*")))
-
-    elif isinstance(seedpoint, (str, PosixPath)):
-        logger.debug("Using this string %s to find datafile(s)", seedpoint)
-        seedpoint_posix = Path(seedpoint)
-        if seedpoint_posix.is_file():
-            logger.debug("%s is file path, will just use this one", seedpoint)
-            datafiles.append(seedpoint)
-    else:
-        logger.debug("%s is list", seedpoint)
-        datafiles.extend(seedpoint)
-    logger.debug("Datafile(s) to use %s", datafiles)
-    return datafiles
-
-
-def export_with_config(config_path, datafile=None, datatype=None):
-    """Export several datatypes with yaml config file
-
-    Args:
-        config_path (str): path to existing yaml file
-        extras (dict): extra arguments
-
-    """
-    logger = logging.getLogger(__file__ + ".export_w_config")
-    logger.debug("Using extras %s", [datafile, datatype])
+    logger = logging.getLogger(__file__ + ".export_tables")
     suffixes = set()
     export_folder = None
     export_path = None
-    try:
-        count = 0
 
-        datafiles, submods, options = read_config(
-            yaml_load(config_path), datafile, datatype
-        )
-        for datafile in datafiles:
-            for submod in submods:
-                logger.info("Exporting %s", submod)
-                export_path = export_results(
-                    datafile,
-                    submod,
-                    config_file=config_path,
-                    **options,
-                )
-                count += 1
-                export_path = Path(export_path)
-                suffixes.add(export_path.suffix)
-        try:
-            export_folder = str(export_path.parent)
-            logger.info("Exported %i files to %s", count, export_folder)
-        except AttributeError:
-            logger.warning("No results exported ")
-    except FileNotFoundError:
-        logger.warning("No config file at: %s", config_path)
+    count = 0
+
+    for datafile in sim2sumoconfig["datafiles"]:
+        for submod in sim2sumoconfig["submods"]:
+            logger.info("Exporting %s", submod)
+            export_path = export_table(
+                datafile,
+                submod,
+                config,
+                **sim2sumoconfig["options"],
+            )
+            count += 1
+            export_path = Path(export_path)
+            suffixes.add(export_path.suffix)
+    try:
+        export_folder = str(export_path.parent)
+        logger.info("Exported %i files to %s", count, export_folder)
+    except AttributeError:
+        logger.warning("No results exported ")
+
     return export_folder, suffixes
 
 
-def parse_args():
-    """Parse arguments for command line tool
-
-    Returns:
-        argparse.NameSpace: the arguments parsed
-    """
-    logger = logging.getLogger(__file__ + ".parse_args")
-    parser = argparse.ArgumentParser(
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-        description="Parsing input to control export of simulator data",
-    )
-
-    parser.add_argument(
-        "--config_path",
-        type=str,
-        help="config that controls the export",
-        default="fmuconfig/output/global_variables.yml",
-    )
-    parser.add_argument(
-        "--env",
-        type=str,
-        help="Which sumo environment to upload to",
-        default="prod",
-    )
-    parser.add_argument(
-        "--datatype",
-        type=str,
-        default=None,
-        help="Override datatype setting, intented for testing only",
-    )
-    parser.add_argument(
-        "--datafile",
-        type=str,
-        default=None,
-        help="Override datafile setting, intented for testing only",
-    )
-    parser.add_argument(
-        "--help_on",
-        type=str,
-        help=(
-            "Use this to get documentation of one of the datatypes to upload\n"
-            + f"valid options are \n{', '.join(SUBMODULES)}"
-        ),
-        default="No help",
-    )
-    parser.add_argument("--d", help="Activate debug mode", action="store_true")
-    args = parser.parse_args()
-    if args.d:
-        logging.basicConfig(
-            level="DEBUG", format="%(name)s - %(levelname)s - %(message)s"
-        )
-    logger.debug("Returning args %s", vars(args))
-    return args
-
-
-def give_help(submod, only_general=False):
-    """Give descriptions of variables available for submodule
-
-    Args:
-        submod (str): submodule
-
-    Returns:
-        str: description of submodule input
-    """
-    general_info = """
-    This utility uses the library ecl2csv, but uploads directly to sumo. Required options are:
-    A config file in yaml format, where you specifiy the variables to extract. What is required
-    is a keyword in the config called "sim2simo". under there you have three optional arguments:
-    * datafile: this can be a string, a list, or it can be absent altogether
-    * datatypes: this needs to be a list, or non existent
-    * options: The options are listed below in the original documentation from ecl2csv. The eclfiles
-               option is replaced with what is under datafile
-
-    """
-    if submod is None:
-        only_general = True
-    if only_general:
-        text_to_return = general_info
-    else:
-        try:
-            text_to_return = general_info + SUBMOD_DICT[submod]["doc"]
-        except KeyError:
-            text_to_return = (
-                f"subtype {submod} does not exist!!, existing options:\n"
-                + "\n".join(SUBMODULES)
-            )
-
-    return text_to_return
-
-
-def upload_with_config(config_path, datafile, datatype, env):
+def upload_tables(sim2sumoconfig, config, env):
     """Upload simulator results to sumo
 
     Args:
@@ -362,15 +173,12 @@ def upload_with_config(config_path, datafile, datatype, env):
         extras (dict): extra arguments
     """
     logger = logging.getLogger(__file__ + ".upload_with_config")
-    logger.debug("Before exporting datafile is %s", datafile)
     logger.debug("Executing with:")
-    logger.debug("config: %s: ", config_path)
+    logger.debug("config keys: %s: ", config.keys())
+    logger.debug("Settings for export_tables %s", sim2sumoconfig)
     logger.debug("Sumo env: %s: ", env)
-
-    upload_folder, suffixes = export_with_config(
-        config_path, datafile, datatype
-    )
-    logger.debug("After exporting datafile is %s", datafile)
+    config_path = config["file_path"]
+    upload_folder, suffixes = export_tables(sim2sumoconfig, config)
     upload(
         upload_folder,
         suffixes,
@@ -378,21 +186,3 @@ def upload_with_config(config_path, datafile, datatype, env):
         env,
         config_path=config_path,
     )
-
-
-def main():
-    """Main function to be called"""
-    logger = logging.getLogger(__file__ + ".main")
-    args = parse_args()
-    logger.debug("Running with arguments %s", args)
-    if args.help_on != "No help":
-        print(give_help(args.help_on))
-    else:
-        logger.info("Will be extracting results")
-        upload_with_config(
-            args.config_path, args.datafile, args.datatype, args.env
-        )
-
-
-if __name__ == "__main__":
-    main()
